functioning handshake
