+++ README.md
@@ -1,1 +1,12 @@
 # GO-Assignment02
+
+a) What are packages in your implementation? What data structure do you use to transmit data and meta-data?
+
+b) Does your implementation use threads or processes? Why is it not realistic to use threads?
+we use threads :)
+
+c) How do you handle message re-ordering?
+
+d) How do you handle message loss?
+
+e) Why is the 3-way handshake important?
